requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiohttp-apigami==0.5.7
aiosignal==1.4.0
apispec==6.9.0
attrs==25.4.0
certifi==2026.1.4
charset-normalizer==3.4.4
colorama==0.4.6
frozenlist==1.8.0
idna==3.11
iniconfig==2.3.0
marshmallow==3.26.2
multidict==6.7.1
packaging==26.0
pluggy==1.6.0
propcache==0.4.1
Pygments==2.19.2
pytest==9.0.2
pytest-aiohttp==1.1.0
pytest-asyncio==1.3.0
PyYAML==6.0.3
requests==2.32.5
urllib3==2.6.3
webargs==8.7.1
yarl==1.22.0
